# Improve memory usage when building large projects