--- a/editor/src/clj/editor/factory.clj
+++ b/editor/src/clj/editor/factory.clj
@@ -14,10 +14,10 @@
 
 (ns editor.factory
   (:require [dynamo.graph :as g]
-            [editor.build-target :as bt]
             [editor.defold-project :as project]
             [editor.graph-util :as gu]
             [editor.outline :as outline]
+            [editor.pipeline :as pipeline]
             [editor.protobuf :as protobuf]
             [editor.protobuf-forms-util :as protobuf-forms-util]
             [editor.resource :as resource]
@@ -67,30 +67,15 @@
       :load-dynamically load-dynamically
       :dynamic-prototype dynamic-prototype)))
 
-(defn build-factory
-  [resource dep-resources user-data]
-  (let [pb-msg (reduce #(assoc %1 (first %2) (second %2))
-                       (:pb-msg user-data)
-                       (map (fn [[label res]] [label (resource/proj-path (get dep-resources res))]) (:dep-resources user-data)))]
-    {:resource resource
-     :content (protobuf/map->bytes (:pb-type user-data) pb-msg)}))
-
-(g/defnk produce-build-targets
-  [_node-id resource factory-type prototype save-value dep-build-targets]
+(g/defnk produce-build-targets [_node-id resource factory-type prototype save-value]
   (or (validation/prop-error :fatal _node-id :prototype validation/prop-nil? prototype "prototype")
-      (let [dep-build-targets (flatten dep-build-targets)
-            deps-by-resource (into {} (map (juxt (comp :resource :resource) :resource) dep-build-targets))
-            dep-resources (map (fn [[label resource]]
-                                 [label (get deps-by-resource resource)])
-                               [[:prototype prototype]])]
-        [(bt/with-content-hash
-           {:node-id _node-id
-            :resource (workspace/make-build-resource resource)
-            :build-fn build-factory
-            :user-data {:pb-msg save-value
-                        :pb-type (get-in factory-types [factory-type :pb-type])
-                        :dep-resources dep-resources}
-            :deps dep-build-targets})])))
+      [(pipeline/make-protobuf-build-target
+         _node-id
+         resource
+         (get-in factory-types [factory-type :pb-type])
+         save-value
+         nil
+         [(resource/proj-path prototype)])]))
 
 (defn load-factory
   [factory-type _project self resource any-factory-desc]
@@ -115,7 +100,6 @@
 (g/defnode FactoryNode
   (inherits resource-node/ResourceNode)
 
-  (input dep-build-targets g/Any)
   (input prototype-resource resource/Resource)
 
   (property factory-type g/Any ; Always assigned in load-fn.
@@ -125,8 +109,7 @@
             (value (gu/passthrough prototype-resource))
             (set (fn [evaluation-context self old-value new-value]
                    (project/resource-setter evaluation-context self old-value new-value
-                                            [:resource :prototype-resource]
-                                            [:build-targets :dep-build-targets])))
+                                            [:resource :prototype-resource])))
             (dynamic error (g/fnk [_node-id prototype-resource]
                                   (or (validation/prop-error :info _node-id :prototype validation/prop-nil? prototype-resource "Prototype")
                                       (validation/prop-error :fatal _node-id :prototype validation/prop-resource-not-exists? prototype-resource "Prototype"))))