… editor/src/clj/editor/pipeline/fontc.clj
… editor/src/clj/editor/sprite.clj
… editor/src/clj/util/digestable.clj
… editor/test/editor/pipeline_test.clj
… editor/test/integration/build_test.clj
… editor/test/integration/sprite_test.clj
… editor/test/integration/test_util.clj
… editor/test/resources/inexact_path_casing_project/.gitignore
… editor/test/resources/inexact_path_casing_project/game.project
… editor/test/resources/inexact_path_casing_project/main/button_cloudy.png
… editor/test/resources/inexact_path_casing_project/main/main.collection
… editor/test/resources/inexact_path_casing_project/main/test.atlas