This change improves:
- memory usage when building large projects (memory reduced by around 15%)
- performance when building large projects (time reduced by around 25%)

Technical notes:

When looking into memory usage during the build, there are several concerns:
- initial memory usage before starting the build. This is not related to the build itself, but it shows the relative importance of the build-time memory usage in relation to normal editor memory usage.
- final memory usage after finishing the build. The difference between final and initial memory usage is the size of the additional in-memory cache data used by the editor.
- peak temporary memory usage during the build. If some phase of the build requires an unexpectedly large amount of memory, it can become a problem when building large projects on machines with limited memory.

In this changeset I mostly focused on allocation hotspots, i.e. places in code that allocate a lot of temporary objects. This changeset mostly fixes low-handing fruits in the build process. This helps a bit with peak temporary memory usage, I assume this is because there is less GC pressure and fewer throw-away objects overwhelming the eden space. As a nice side-effect, reducing temporary allocations makes the build process noticeably faster.

I measured 3 "modes" of the build process:
- "full": when the project is built, there is no on-disc or in-memory cache.
- "cold": when the project is built, there is an on-disc cache but the in-memory cache is empty (i.e. build after starting the editor).
- "hot": when the project is built, there is both an on-disc and in-memory cache (i.e. build after a previous build).

All measurements were done after running GC. Typical initial memory usage is between **8 and 10 GB** before starting the build. After the build, the editor memory usage increases by around **4 to 5** GB.

Here is the before/after comparison for this changeset:

1. Peak temporary memory usage (measured as peak memory usage during the build minus final memory usage after the build):
   mode | before    | after    | improvement
   -----|----------:|---------:|-----------:
   full | `11.2 GB` | `7.7 GB` | `30.8%`
   cold | `8.2 GB`  | `5.5 GB` | `33.2%`
   hot  | `11.3 GB` | `6.6 GB` | `41.7%`

   How is this possible that hot build actually uses more throw-away memory than the cold one, if it does less by definition? Great question. We do dynamic build target resolution, which involves creating a lot of throw-away data structures. Probably GC is invoked less frequently when the build only does that... The numbers should probably be taken with a grain of salt. Still, the improvement is observable.
2. Build time:
   mode | before   | after   | improvement
   -----|---------:|--------:|-----------:
   full | `12m30s` | `8m20s` | `33.3%`
   cold | `10m15s` | `6m11s` | `39.7%`
   hot  |    `28s` |   `25s` | `8.6%`

   I observed some variation between measurements, e.g. +- 30s time per cold build is typical. Even with the variation, the improvement is noticeable.
3. Average temporary memory usage (measured as average memory usage throughout the build process minus final memory usage after the build). Average memory was calculated by collecting used memory every second during the build. The idea of this metric is to measure GC pressure: if we create fewer temporary objects, there saw pattern of memory usage is less spikey, and that should make the average memory lower.
   mode | before    | after    | improvement
   -----|----------:|---------:|-----------:
   full | `1.9 GB`  | `1.4 GB` | `26.8%`
   cold | `1.2 GB`  | `0.4 GB` | `65.7%`
   hot  | `3.1 GB`  | `2.6 GB` | `16.1%`

   Yet again, somehow the hot build uses more temporary memory than the cold one. I think this metric is not the most useful, because it smoothes out the peaks, and if there are particular points in the build process that are problematic, they will not be visible in this metric. Still, reducing the amount of allocations in hot loops is visible here.

Future work:
When investigating, I found some places that generate very big peaks in memory usage (up to 10 GBs!) during the build. Fixing those is more involved, and I will address them in a separate PR. The problematic places are:
- clj-antlr (used in `editor.lua-parser/lua-info` and `lua-parser/parse-requires`) creates a lot of lazy sequences
- `texture-page-counts` output of `GuiSceneNode` creates a lot of red/black tree nodes

Related to #10113